bug 1407806: Move tests to test_views_code.py
